level added

+++ client/src/components/CreateCourse/CreateCourse.jsx
@@ -4,24 +4,18 @@ import SelectImage from "../../images/selectImage.png";
 import axios from "axios";
 import { CircularProgress } from "@mui/material";
 import { useDispatch, useSelector } from "react-redux";
-import { attachedFilesActions } from "../../store/attachedFiles";
 import { choicesActions } from "../../store/choices";
 import { questionsActions } from "../../store/questions";
-import UploadedFile from "../UploadedFile/UploadedFile";
+import { levelsActions } from "../../store/levels";
 import { useNavigate } from "react-router-dom";
 import Choice from "../Choice/Choice";
+import Level from "../Level/Level";
 import Question from "../Question/Question";
 
 const CreateCourse = ({ token }) => {
   const navigate = useNavigate();
 
   const dispatch = useDispatch();
-  const attachedFiles = useSelector(
-    (state) => state.attachedFiles.attachedFiles
-  );
-  const attachedFilesNum = useSelector(
-    (state) => state.attachedFiles.attachedFilesNum
-  );
 
   const choices = useSelector((state) => state.choices.choices);
   const choicesNum = useSelector((state) => state.choices.choicesNum);
@@ -29,17 +23,24 @@ const CreateCourse = ({ token }) => {
   const questions = useSelector((state) => state.questions.questions);
   const questionsNum = useSelector((state) => state.questions.questionsNum);
 
+  const levels = useSelector((state) => state.levels.levels);
+  const levelsNum = useSelector((state) => state.levels.levelsNum);
+
   const [courseId, setCourseId] = useState("");
   const [courseName, setCourseName] = useState("");
   const [courseGoal, setCourseGoal] = useState("");
   const [courseBio, setCourseBio] = useState("");
   const [courseImageDisplay, setCourseImageDisplay] = useState(SelectImage);
   const [courseImage, setCourseImage] = useState(null);
+  const [levelImageDisplay, setLevelImageDisplay] = useState(SelectImage);
+  const [levelImage, setLeveleImage] = useState(null);
+  const [levelTitle, setLeveleTitle] = useState("");
+  const [levelDescription, setLeveleDescription] = useState("");
 
   const [courseCreated, setCourseCreated] = useState(false);
-  const [docsAdded, setDocsAdded] = useState(false);
   const [isCreating, setIsCreating] = useState(false);
   const [filesUploading, setFilesUploading] = useState(false);
+  const [jobDone, setJobDone] = useState(false);
 
   const [question, setQuestion] = useState("");
   const [answer, setAnswer] = useState("");
@@ -80,45 +81,48 @@ const CreateCourse = ({ token }) => {
       console.log(res);
       setCourseId(res.data.id);
       setIsCreating(false);
-      dispatch(attachedFilesActions.deleteAllAttachedFiles());
+      dispatch(choicesActions.deleteAllChoices());
+      dispatch(questionsActions.deleteAllQuestions());
       setCourseCreated(true);
     } catch (err) {
       console.log(err);
     }
   };
 
-  const uploadDoc = async (e) => {
-    setFilesUploading(true);
-    console.log(e.target.files);
-    console.log(Array.from(e.target.files || []));
-    const filesList = Array.from(e.target.files || []);
-
-    for (let i = 0; i < filesList.length; i++) {
-      dispatch(attachedFilesActions.addAttachedFiles(filesList[i]));
-      const formData = new FormData();
-      formData.append("avatar", filesList[i]);
-      const res = await axios.post(
-        `${process.env.REACT_APP_API_ADDRESS}uploadedFiles/addDoc/${courseId}`,
-        formData,
-        {
-          headers: {
-            token: `Bearer ${token}`,
-          },
-        }
+  const createLevel = () => {
+    console.log(questions);
+    if (
+      levelTitle !== "" &&
+      levelImage !== null &&
+      questionsNum !== 0 &&
+      levelDescription !== ""
+    ) {
+      dispatch(
+        levelsActions.addLevel({
+          title: levelTitle,
+          doc: levelImage,
+          desc: levelDescription,
+          questions: questions,
+        })
       );
-      console.log(res);
+      dispatch(questionsActions.deleteAllQuestions());
+      dispatch(choicesActions.deleteAllChoices());
+      setChoice("");
+      setQuestion("");
+      setAnswer("");
+      setLeveleImage(null);
+      setLevelImageDisplay(SelectImage);
+      setLeveleTitle("");
+      setLeveleDescription("");
     }
-
-    setFilesUploading(false);
-    console.log(attachedFiles);
-
-    console.log(attachedFiles);
   };
 
-  const finishAddingFiles = async () => {
-    setDocsAdded(true);
-    dispatch(choicesActions.deleteAllChoices());
-    dispatch(questionsActions.deleteAllQuestions());
+  const handleUploadedLevelDoc = (e) => {
+    setFilesUploading(true);
+    console.log(e.target.files[0]);
+    setLeveleImage(e.target.files[0]);
+    setLevelImageDisplay(URL.createObjectURL(e.target.files[0]));
+    setFilesUploading(false);
   };
 
   const addChoice = () => {
@@ -129,28 +133,31 @@ const CreateCourse = ({ token }) => {
   };
 
   const finishCreatingCourse = async () => {
-    console.log(courseId);
-    console.log(questions);
-    console.log(questionsNum);
-    for (let i = 0; i < questionsNum; i++) {
-      await axios.post(
-        `${process.env.REACT_APP_API_ADDRESS}questions/addQuestion/${courseId}`,
-        {
-          question: questions[i].question,
-          fullAnswer: questions[i].fullAnswer,
-          choices: questions[i].choices,
-        },
-        {
-          headers: {
-            token: `Bearer ${token}`,
-            "Content-Type": "application/json",
-          },
-        }
-      );
+    console.log(levels);
+    try {
+      setJobDone(true);
+      for (let i = 0; i < levelsNum; i++) {
+        const formData = new FormData();
+        formData.append("title", levels[i].title);
+        formData.append("description", levels[i].desc);
+        formData.append("avatar", levels[i].doc);
+        formData.append("questions", JSON.stringify(levels[i].questions));
+
+        await axios.post(
+          `${process.env.REACT_APP_API_ADDRESS}levels/addLevel/${courseId}`,
+          formData,
+          {
+            headers: {
+              token: `Bearer ${token}`,
+            },
+          }
+        );
+      }
+      navigate("/profileStructure/myCourses");
+      setJobDone(false);
+    } catch (err) {
+      console.log(err);
     }
-    dispatch(choicesActions.deleteAllChoices());
-    dispatch(questionsActions.deleteAllQuestions());
-    navigate("/profileStructure/myCourses");
   };
 
   const addQuestion = () => {
@@ -196,7 +203,6 @@ const CreateCourse = ({ token }) => {
                 setCourseName(e.target.value);
               }}
               required
-              readOnly={courseCreated}
             />
             <span>Title</span>
           </div>
@@ -207,7 +213,6 @@ const CreateCourse = ({ token }) => {
                 setCourseGoal(e.target.value);
               }}
               required
-              readOnly={courseCreated}
             />
             <span>Goals</span>
           </div>
@@ -217,7 +222,6 @@ const CreateCourse = ({ token }) => {
                 setCourseBio(e.target.value);
               }}
               required
-              readOnly={courseCreated}
             />
             <span>Abstract</span>
           </div>
@@ -230,70 +234,63 @@ const CreateCourse = ({ token }) => {
           </button>
         )}
       </div>
-
       <div
-        className={`${classes.courseDoc} ${
-          courseCreated && classes.nowAddDoc
-        } ${docsAdded && classes.finishAddingFiles}`}
-      >
-        <label
-          for="uploadDoc"
-          className={`${classes.addDoc}  ${
-            filesUploading && classes.filesUploading
-          }`}
-        >
-          Add New Document
-          <input
-            id="uploadDoc"
-            type="file"
-            multiple="multiple"
-            accept=".pptx, .doc, .pdf, .jpg, .png, .jpeg"
-            onChange={uploadDoc}
-            disabled={filesUploading}
-          />
-        </label>
-        <div
-          className={`${classes.addedDocs} ${
-            filesUploading && classes.filesUploading
-          }`}
-        >
-          {attachedFilesNum !== 0 && (
-            <div className={classes.uploadedFilesList}>
-              {attachedFiles.map((item, index) => (
-                <UploadedFile
-                  key={index}
-                  token={token}
-                  id={item.id}
-                  type={item.name.split(".")[item.name.split(".").length - 1]}
-                  name={item.name}
-                  downloadOrDelete="delete"
-                />
-              ))}
-            </div>
-          )}
-        </div>
-        {filesUploading ? (
-          <CircularProgress />
-        ) : (
-          <div className={classes.finishAddingFiles}>
-            <button onClick={finishAddingFiles}>Finish Adding Files :)</button>
-          </div>
-        )}
-      </div>
-
-      <div
-        className={`${classes.addQuestion} ${
-          docsAdded && classes.nowAddQuestion
+        className={`${classes.addLevel} ${
+          courseCreated && classes.nowAddLevel
         }`}
       >
+        <div className={classes.levelDetails}>
+          <h3>Now Create Some Levels for this Course</h3>
+          <div className={classes.enterData}>
+            <input
+              value={levelTitle}
+              type="text"
+              onChange={(e) => {
+                setLeveleTitle(e.target.value);
+              }}
+              required
+            />
+            <span>Level Title</span>
+          </div>
+          <div className={classes.enterData}>
+            <textarea
+              value={levelDescription}
+              type="text"
+              onChange={(e) => {
+                setLeveleDescription(e.target.value);
+              }}
+              required
+            />
+            <span>Level Desciption</span>
+          </div>
+          <label
+            for="uploadDoc"
+            className={`${classes.addDoc}  ${
+              filesUploading && classes.filesUploading
+            }`}
+          >
+            Add New Document
+            <input
+              id="uploadDoc"
+              type="file"
+              multiple="multiple"
+              accept=".jpg, .png, .jpeg, .jfif"
+              onChange={handleUploadedLevelDoc}
+              disabled={filesUploading}
+            />
+          </label>
+          <div className={classes.uploadedDoc}>
+            <img src={levelImageDisplay} alt="Level_Image" />
+          </div>
+        </div>
         <div className={classes.questionForm}>
+          <h3>Add Qusetions for this Level</h3>
           <div className={classes.questionData}>
             <input
               type="text"
               value={question}
               onChange={(e) => setQuestion(e.target.value)}
               required
-              readOnly={courseCreated}
             />
             <span>Question</span>
           </div>
@@ -303,22 +300,22 @@ const CreateCourse = ({ token }) => {
               value={answer}
               onChange={(e) => setAnswer(e.target.value)}
               required
-              readOnly={courseCreated}
             />
             <span>Ful Answer</span>
           </div>
           <div className={classes.choiceData}>
-            <textarea
-              value={choice}
-              onChange={(e) => setChoice(e.target.value)}
-            />
+            <input value={choice} onChange={(e) => setChoice(e.target.value)} />
             <button className={classes.addChoice} onClick={addChoice}>
               Add Choice
             </button>
             {choicesNum !== 0 && (
               <div className={classes.choices}>
                 {choices.map((item, index) => (
-                  <Choice answer={item.choice} setChoice={setChoice} />
+                  <Choice
+                    key={index}
+                    answer={item.choice}
+                    setChoice={setChoice}
+                  />
                 ))}
               </div>
             )}
@@ -327,25 +324,55 @@ const CreateCourse = ({ token }) => {
             Creating Question Finished :)
           </button>
         </div>
-        <div className={classes.questions}>
-          {questions.map((item, index) => (
-            <Question
-              question={item.question}
-              fullAnswer={item.fullAnswer}
-              choices={item.choices}
-              setAnswer={setAnswer}
-              setQuestion={setQuestion}
-            />
-          ))}
-        </div>
+        {questionsNum !== 0 ? (
+          <div className={classes.questions}>
+            {questions.map((item, index) => (
+              <Question
+                key={index}
+                question={item.question}
+                fullAnswer={item.fullAnswer}
+                choices={item.choices}
+                setAnswer={setAnswer}
+                setQuestion={setQuestion}
+              />
+            ))}
+          </div>
+        ) : (
+          <p className={classes.empty}>No Questions</p>
+        )}
+        <button className={classes.finishCreatingLevel} onClick={createLevel}>
+          Creating Level Finished :)
+        </button>
+        {levelsNum !== 0 ? (
+          <div className={classes.levels}>
+            {levels.map((item, index) => (
+              <Level
+                key={index}
+                title={item.title}
+                doc={item.doc}
+                desc={item.desc}
+                questions={item.questions}
+                setTitle={setLeveleTitle}
+                setDoc={setLeveleImage}
+                setDocDisplay={setLevelImageDisplay}
+                setDesc={setLeveleDescription}
+              />
+            ))}
+          </div>
+        ) : (
+          <p className={classes.empty}>No Levels</p>
+        )}
       </div>
-
-      <button
-        className={classes.finishCreatingCourse}
-        onClick={finishCreatingCourse}
-      >
-        Finish Creating Course :))
-      </button>
+      {!jobDone ? (
+        <button
+          className={classes.finishCreatingCourse}
+          onClick={finishCreatingCourse}
+        >
+          Finish Creating Course :))
+        </button>
+      ) : (
+        <CircularProgress />
+      )}
     </div>
   );
 };
